+++ lib/simple-lib-spring-opensearch-appender-bulk-only-1.0.0/src/main/java/com/cube/simple/opensearch/config/OpenSearchProperties.java
@@ -35,5 +35,4 @@ public class OpenSearchProperties {
     public void addOpenSearchProperty(Property property) {
         properties.add(property);
     }
-
 }